edge/face-recognition/requirements.txt
face_recognition_models
Click
dlib
numpy
Pillow
scipy
picamera[array]
AWSIoTPythonSDK
scikit-learn
opencv-python
isodate
boto3
requests
ffmpeg-python
